pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "bulla"
version = "0.36.0"
description = "Witness kernel for agentic compositions — diagnose, attest, seal"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.10"
authors = [{ name = "Glyph" }]
keywords = ["mcp", "agents", "witness", "bulla", "composition", "coherence"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: Other/Proprietary License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Quality Assurance",
]
dependencies = [
    "pyyaml>=6.0",
    # tomllib is stdlib on 3.11+; tomli is the canonical backport for 3.10.
    "tomli>=2.0; python_version < '3.11'",
]

[project.optional-dependencies]
ots = ["opentimestamps-client>=0.7"]
discover = ["openai>=1.0", "anthropic>=0.20"]
# Framework adapters — optional. Static AST parsing has no runtime dep on
# the framework itself; these extras enable schema-validation helpers and
# reserve the dependency surface for future ParseMode.RUNTIME support.
langgraph = ["langchain-core>=0.3"]
crewai = ["crewai>=0.80"]
all = ["bulla[ots,discover,langgraph,crewai]"]

[project.scripts]
bulla = "bulla.cli:main"

[project.urls]
Homepage = "https://github.com/jkomkov/bulla"
Documentation = "https://github.com/jkomkov/bulla/blob/main/WITNESS-CONTRACT.md"
Issues = "https://github.com/jkomkov/bulla/issues"
Changelog = "https://github.com/jkomkov/bulla/blob/main/CHANGELOG.md"

[tool.hatch.build.targets.wheel]
packages = ["src/bulla"]

[tool.pytest.ini_options]
testpaths = ["tests"]
